requirements.txt
numpy==1.26.4
PyPDF2==3.0.1
mistralai==0.4.0
faiss-cpu==1.8.0
streamlit==1.35.0
streamlit-extras==0.4.3